refactor: totalPrice, checkBox관련 함수들 customhook으로 분리
